// src/generic.d.ts
export type ForwardRefComponent<P, T> = React.ForwardRefExoticComponent<P & React.RefAttributes<T>>
// ...
export type SemanticShorthandItemFunc<TProps> = (
  component: React.ElementType<TProps>,
  props: TProps,
  children?: React.ReactNode,
) => React.ReactElement<any> | null

export type ShorthandRenderFunction<C extends React.ElementType, P> = (
  Component: C,
  props: P,
) => React.ReactNode

export type SemanticShorthandCollection<TProps extends Record<string, any>> =
  SemanticShorthandItem<TProps>[]
export type SemanticShorthandContent = React.ReactNode
export type SemanticShorthandItem<TProps extends Record<string, any>> =
  | React.ReactNode
  | SemanticShorthandItemFunc<TProps>
  | (Omit<TProps, 'children'> & {
      // Not all TProps can have `children`, without this condition it will match to "any"
      children?: TProps extends { children: any }
        ? TProps['children'] | ShorthandRenderFunction<React.ElementType<TProps>, TProps>
        : ShorthandRenderFunction<React.ElementType<TProps>, TProps>
    })